refactor: remove deprecated v1 api for member
